requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.5.1
attrs==25.1.0
beautifulsoup4==4.12.3
bitarray==3.0.0
black==24.8.0
boto3==1.35.61
botocore==1.35.61
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.0
cryptography==44.0.2
cssselect==1.2.0
cycler==0.12.1
dask==2024.12.0
docker-py==1.10.6
docker-pycreds==0.4.0
et_xmlfile==2.0.0
fonttools==4.54.1
fsspec==2024.10.0
future==1.0.0
h11==0.16.0
halo==0.0.31
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
impyla==0.19.0
iniconfig==2.0.0
jmespath==1.0.1
JPype1==1.5.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
locket==1.0.0
log-symbols==0.0.14
lxml==5.3.0
matplotlib==3.9.2
mcp==1.12.0
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
mysql_mcp_server==0.2.2
numpy==2.1.3
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parquet-cli==1.3
parquet_tools==0.2.16
partd==1.4.2
pathspec==0.12.1
pillow==11.0.0
pipx==1.7.1
platformdirs==4.3.6
pluggy==1.5.0
protobuf==3.20.3
pure-sasl==0.6.2
py4j==0.10.9.7
pyaml==24.9.0
pyarrow==18.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyHive==0.7.0
pyparsing==3.2.0
pyquery==2.0.1
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2024.2
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.26.0
s3transfer==0.10.3
seaborn==0.13.2
selenium==4.28.1
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spinners==0.0.24
sse-starlette==2.4.1
starlette==0.47.2
tabulate==0.9.0
termcolor==2.5.0
thrift==0.16.0
thrift-sasl==0.4.3
toolz==1.0.0
trio==0.29.0
trio-websocket==0.12.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
userpath==1.9.2
uv==0.8.0
uvicorn==0.35.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0